Removed comments and sped up BMP
All redundant text has been removed to reduce the module size. Only the
required parameters of the BMP are stored, and instead of pixel by
pixel transmission, 64 pixels are sent in one go. The colour order had
to be updated form BGR to RGB, which slows it down still. But that's
pretty much it. Also, small changes can cause strange Memory Errors?!